requirements.txt
annotated-types==0.7.0
anyio==4.4.0
azure-core==1.31.0
azure-storage-blob==12.22.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
distro==1.9.0
Flask==3.0.3
Flask-Uploads==0.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.9
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.45.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
python-dotenv

